+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # 🔖 Changelog
 
 ## Unreleased
+
+Features:
+
+- Implement the `IsValidFirestoreId` decorator.